% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_urls.R
\name{crawl_urls}
\alias{crawl_urls}
\title{Lookup DOIs from publisher URLs}
\usage{
crawl_urls(urls)
}
\arguments{
\item{url}{A vector of strings corresponding to the publisher or other
URL for an article that does not contain a doi within the link.}
}
\value{
A vector of character strings corresponding to the DOIs.
}
\description{
Takes a vector of publisher URLs for an article, downloads
the page code and extracts the first doi URLs from each page (i.e. a URL
that contains 'doi.org').
}
\examples{
\dontrun{
urls <- c('https://www.sciencedirect.com/science/article/pii/S0048969703000068',
         'https://www.cambridge.org/core/journals/aquatic-living-resources/article/differences-in-abiotic-water-conditions-between-fluvial-reaches-and-crayfish-fauna-in-some-northern-rivers-of-the-iberian-peninsula/94940FBD4D49B6E36857CD074DA72FB6',
         'https://www.publish.csiro.au/MF/MF04111')
dois <- crawl_urls(url)
}
}
