% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abstract.R
\name{get_abstract}
\alias{get_abstract}
\title{Get abstracts for bibliographic records missing them}
\usage{
get_abstract(bib_data)
}
\arguments{
\item{bib_data}{A bibliographic data set stored as a dataframe, with
each record as a row. Records must contains a column called 'abstract'
that contains the record title, and a column called 'doi' that
contains (or would contain) the DOIs.}
}
\value{
A revised dataframe containing records for which the abstract
could be retrieved.
}
\description{
Retrieve abstracts for records that are missing them.
The function uses the crossref API matching on DOI (digital object
identifier).
}
\examples{
\dontrun{
bib_data <- synthesisr::read_refs('inst/extdata/export.ris')
bib_data_new <- get_abstract(bib_data)
}
}
