% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_doi.R
\name{get_doi}
\alias{get_doi}
\title{Get digital object identifiers}
\usage{
get_doi(bib_data)
}
\arguments{
\item{bib_data}{A bibliographic data set stored as a dataframe, with
each record as a row. Records must contains a column called 'title'
that contains the record title, and a column called 'doi' that
contains (or would contain) the DOIs.}
}
\value{
A revised dataframe containing DOIs for records for which
the DOI could be retrieved.
}
\description{
Retrieve digital object identifiers (DOIs) for records
that are missing them. The function
}
\examples{
\dontrun{
bib_data<-revtools::read_bibliography('inst/extdata/export.ris')
}
}
