% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal_abbreviation.R
\name{journal_abbreviation}
\alias{journal_abbreviation}
\title{journal_abbreviation}
\usage{
journal_abbreviation(data, journal_names)
}
\arguments{
\item{data}{A dataframe}

\item{journal_names}{column in the dataframe that contains the titles of the sources}
}
\description{
convert between abbreviated and long forms of journal names based on an updatable dataframe of abbreviations and full names.
Support matching where journal cannot be found but there is an almost full match with abbreviated or unabbreviated journal name
}
\examples{
\dontrun{
bib_data <- synthesisr::read_refs('inst/extdata/export.ris')
bib_data_new <- journal_abbreviation(bib_data, bib_data$journal)
}

}
