% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_health.R
\name{plot_health}
\alias{plot_health}
\title{Plot RIS file health
remotes::install_git("https://git.rud.is/hrbrmstr/ggchicklet.git")
report <- scan_file(refs)
health_plot <- plot_health(report)
health_plot}
\usage{
plot_health(health_check, session = NULL)
}
\description{
Plot RIS file health
remotes::install_git("https://git.rud.is/hrbrmstr/ggchicklet.git")
report <- scan_file(refs)
health_plot <- plot_health(report)
health_plot
}
