% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_abstracts.R
\name{remove_abstracts}
\alias{remove_abstracts}
\title{Remove abstracts from an RIS file}
\usage{
remove_abstracts(
  input_filename,
  save = TRUE,
  output_filename = "export",
  path = NULL
)
}
\arguments{
\item{bib_data}{A bibliographic data set stored as an RIS file.}
}
\value{
A revised RIS file containing records minus abstracts.
}
\description{
Strip out abstracts, which may be copyrighted, from an
RIS file.
}
\examples{
\dontrun{
bib_data <- remove_abstracts('inst/extdata/data.csv')
}
}
