% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBibfixApp.R
\name{runBibfixApp}
\alias{runBibfixApp}
\title{Run Shiny App}
\usage{
runBibfixApp(app_name = "bibfix")
}
\arguments{
\item{app_name}{bibfix}
}
\value{
Launches the Shiny application.
}
\description{
This function launches a Shiny application from the specified directory in an R package.
}
