% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_file.R
\name{scan_file}
\alias{scan_file}
\title{Scan RIS file for damage}
\usage{
scan_file(refs)
}
\description{
Scan RIS file for damage
}
\examples{
\dontrun{
library(openalex)
devtools::install_github("kth-library/openalex", dependencies = TRUE)
openalex_polite("neal_haddaway@hotmail.com")
library(openalexR)
file <- "Users/neal.haddaway/OneDrive - SEI/ESHackathon/4.Remote 2020/citationchaser private/archive/inst/shiny-examples/citationchaser/www/references.ris"
refs <- synthesisr::read_refs(file.choose())
report <- scan_file(refs)
}
}
