% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_openAlex.R
\name{search_openAlex}
\alias{search_openAlex}
\title{Search Open Alex API for records matching different identifiers}
\usage{
search_openAlex(input)
}
\arguments{
\item{input}{Two column dataframe consisting of one column named 'ids'
contains the identifiers, including 'doi' (digital object identifier),
'mag' (microsoft academic identifier), 'openalex' (Open Alex
identifier), 'pmid' (PubMed identifier), 'pmcid (PubMed Central
identifier), and 'title' (record title), and one column that contains
the corresponding identifier type for each ID (using those names listed
here).}
}
\value{
a list of character vectors of named reference items returned
from the Open Alex API.
}
\description{
Search the Open Alex API to bring back records matching
one or more identifiers. Searches based on identifier OR title
(separate query type needed for each).
}
\examples{
\dontrun{
devtools::install_github("kth-library/openalex", dependencies = TRUE)
library(openalex)
openalex_polite("neal_haddaway@hotmail.com")
input <- data.frame(ids = c('10.1371/journal.pone.0138237', '10.1186/s13750-016-0059-6', 'The role of tourism and recreation in the spread of non-native species: a systematic review and meta-analysis'),
                    type = c('doi', 'doi', 'title'))
results <- search_openAlex(input)
}
}
