% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wos_abbrev_table}
\alias{wos_abbrev_table}
\title{wos_abbrev_table}
\format{
A data frame with 87224 rows and 3 variables:
\describe{
\item{\code{full}}{character Full name of the Journal}
\item{\code{abbrev}}{character Abbreviation of the Journal}
\item{\code{abbrev.dots}}{Abbreviation of the Journal formatted with . after each word}
}
}
\usage{
wos_abbrev_table
}
\description{
A list of the Web of Science Journal abbreviations
}
\details{
DETAILS
}
\keyword{datasets}
