% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urls.R
\name{get_urls}
\alias{get_urls}
\title{Get URL links}
\usage{
get_urls(bib_data)
}
\arguments{
\item{bib_data}{A bibliographic data set stored as a dataframe, with
each record as a row. Records must contains a column called 'doi' that
contains the DOIs for the records.}
}
\value{
A revised dataframe containing various urls for records that
could be found using the API.
}
\description{
Retrieve URL links to the record in various formats and
from various sources.
}
\examples{
\dontrun{
bib_data<-revtools::read_bibliography('inst/extdata/export.ris')
}
}
