% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_health.R
\name{plot_health}
\alias{plot_health}
\title{Plot RIS file health
remotes::install_git("https://git.rud.is/hrbrmstr/ggchicklet.git")
health_check <- scan_file(refs)}
\usage{
plot_health(health_check)
}
\description{
Plot RIS file health
remotes::install_git("https://git.rud.is/hrbrmstr/ggchicklet.git")
health_check <- scan_file(refs)
}
