% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_health.R
\name{plot_health}
\alias{plot_health}
\title{Plot RIS file health}
\usage{
plot_health(health_check, session = NULL)
}
\arguments{
\item{health_check}{result of bibfix::scan_file()}

\item{session}{if used in shiny the plot is a plotly version}
}
\description{
Plot RIS file health
}
