% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_blanks.R
\name{repair_refs}
\alias{repair_refs}
\title{Find empty fields across bibliographic data}
\usage{
repair_refs(refs, replace_abstracts = FALSE)
}
\arguments{
\item{refs}{A data frame outputted from synthesisr.}
}
\value{
A data frame matching the input in style, but containing filled in
information where there were blanks. Where multiple matches were obtained on
Open Alex, these are provided for the user to decide which to retain.
}
\description{
Fill in missing information for a bibliuographic dataset by
searching the Open Alex API.
}
\examples{
\dontrun{
library(openalex)
devtools::install_github("kth-library/openalex", dependencies = TRUE)
openalex_polite("neal_haddaway@hotmail.com")
devtools::install_github("massimoaria/openalexR", dependencies = TRUE)
library(openalexR)
file <- "Users/neal.haddaway/OneDrive - SEI/ESHackathon/4.Remote 2020/citationchaser private/archive/inst/shiny-examples/citationchaser/www/references.ris"
refs <- synthesisr::read_refs(file.choose())
repaired_refs <- repair_refs(refs, replace_abstracts = TRUE)
}
}
