% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_file.R
\name{scan_file}
\alias{scan_file}
\title{Scan RIS file for damage}
\usage{
scan_file(refs)
}
\arguments{
\item{refs}{A RIS file with biblographic records}
}
\description{
Scan RIS file for damage
}
\examples{
\dontrun{
library(openalexR)
refs <- synthesisr::read_refs(file.choose())
report <- scan_file(refs)
}
}
